% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmcCustomMultiModalSpeckle}
\alias{mcmcCustomMultiModalSpeckle}
\title{Generate Multimodal Speckle Pattern using MCMC}
\usage{
mcmcCustomMultiModalSpeckle(rows, cols, means, std_devs, weights, iterations)
}
\arguments{
\item{rows}{The number of rows in the output matrix.}

\item{cols}{The number of columns in the output matrix.}

\item{means}{A NumericVector of means for the Gaussian distributions.}

\item{std_devs}{A NumericVector of standard deviations for the Gaussian distributions.}

\item{weights}{A NumericVector of weights for the Gaussian distributions.}

\item{iterations}{The number of MCMC iterations.}
}
\value{
A NumericMatrix representing the speckle pattern.
}
\description{
This function generates a multimodal speckle pattern using Markov Chain Monte Carlo (MCMC) methods.
}
\examples{
\dontrun{
rows <- 256
cols <- 256
means <- c(0.3, 0.7)
std_devs <- c(0.1, 0.1)
weights <- c(0.5, 0.5)
iterations <- 500 
speckle <- mcmcCustomMultiModalSpeckle(rows, cols, means, std_devs, weights, iterations)
image(speckle, col = gray.colors(256), main = "Custom Multi-Modal Speckle")
}
}
