\name{SC24002027-package}
\alias{SC24002027-package}
\alias{SC24002027}
\docType{package}
\title{
  This R package is used to calculate the reconstructed image uncertainty using bootstrap method and to generate bimodal Gaussian distributed speckle using MCMC method
}
\description{
  Several R and Rcpp function are included in this package for illustration. 
}
\author{
Wang Peiyi

Maintainer: Wang Peiyi <1291419649@qq.com>
}
\keyword{ package }
