% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{targetDensity}
\alias{targetDensity}
\title{Compute Target Density for Multimodal Distribution}
\usage{
targetDensity(x, means, std_devs, weights)
}
\arguments{
\item{x}{The value at which to calculate the density.}

\item{means}{A NumericVector of means for the Gaussian distributions.}

\item{std_devs}{A NumericVector of standard deviations for the Gaussian distributions.}

\item{weights}{A NumericVector of weights for the Gaussian distributions.}
}
\value{
The computed target density value.
}
\description{
This function computes the target density for a given value x,
}
